# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GenerateStoryRequestInterestsItem =
  (typeof GenerateStoryRequestInterestsItem)[keyof typeof GenerateStoryRequestInterestsItem];

export const GenerateStoryRequestInterestsItem = {
  dinosaurs: "dinosaurs",
  space: "space",
  princess: "princess",
  animals: "animals",
  cars: "cars",
  magic: "magic",
} as const;

/**
 * Desired reading length of the story
 */
export type GenerateStoryRequestStoryLength =
  (typeof GenerateStoryRequestStoryLength)[keyof typeof GenerateStoryRequestStoryLength];

export const GenerateStoryRequestStoryLength = {
  "5min": "5min",
  "10min": "10min",
  "15min": "15min",
} as const;

/**
 * Tone of the story
 */
export type GenerateStoryRequestTone =
  (typeof GenerateStoryRequestTone)[keyof typeof GenerateStoryRequestTone];

export const GenerateStoryRequestTone = {
  calm: "calm",
  exciting: "exciting",
  silly: "silly",
  adventurous: "adventurous",
  magical: "magical",
} as const;

export interface GenerateStoryRequest {
  /** The child's name */
  childName: string;
  /**
   * The child's age
   * @minimum 1
   * @maximum 12
   */
  age: number;
  /** The child's interests */
  interests: GenerateStoryRequestInterestsItem[];
  /** Desired reading length of the story */
  storyLength?: GenerateStoryRequestStoryLength;
  /** Tone of the story */
  tone?: GenerateStoryRequestTone;
  /** Child profile ID for memory lookup */
  childId?: number;
  /** Series ID for memory lookup */
  seriesId?: number;
}

export interface GenerateStoryResponse {
  /** The story title */
  title: string;
  /** The full story text */
  story: string;
  /** A fun emoji that represents the story */
  emoji: string;
  /** Short 3-5 sentence summary of the story */
  summary: string;
}

export interface SaveStoryRequest {
  childName: string;
  emoji: string;
  title: string;
  story: string;
  /** Short 3-5 sentence summary for use in next episode */
  storySummary?: string;
  /** Comma-separated interests */
  interests: string;
  /** ID of the child profile to link this story to */
  childId?: number;
  /** ID of the series this story belongs to */
  seriesId?: number;
  /** Episode number within the series */
  episodeNumber?: number;
}

export interface SavedStory {
  id: number;
  childName: string;
  emoji: string;
  title: string;
  story: string;
  /**
   * Short summary stored with the story
   * @nullable
   */
  storySummary?: string | null;
  interests: string;
  createdAt: string;
  /** @nullable */
  childId?: number | null;
  /** @nullable */
  seriesId?: number | null;
  /** @nullable */
  episodeNumber?: number | null;
}

export interface MemoryCharacter {
  name: string;
  description?: string;
}

export interface MemoryLocation {
  name: string;
  description?: string;
}

export interface StoryMemory {
  id: number;
  userId: string;
  childId: number;
  /** @nullable */
  seriesId?: number | null;
  /** Main recurring characters */
  mainCharacters: MemoryCharacter[];
  /** Supporting characters */
  sideCharacters: MemoryCharacter[];
  /** Settings and locations */
  locations: MemoryLocation[];
  /** Recurring story themes */
  themes: string[];
  /** Observed tone preferences */
  tonePreferences: string[];
  updatedAt: string;
}

export interface UpsertMemoryRequest {
  childId: number;
  /** @nullable */
  seriesId?: number | null;
  mainCharacters?: MemoryCharacter[];
  sideCharacters?: MemoryCharacter[];
  locations?: MemoryLocation[];
  themes?: string[];
  tonePreferences?: string[];
}

export interface StorySeries {
  id: number;
  userId: string;
  childId: number;
  title: string;
  /** @nullable */
  theme?: string | null;
  /** Number of stories in this series */
  storyCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSeriesRequest {
  childId: number;
  title: string;
  theme?: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface StreakData {
  /** Current streak count (0 if no activity yet) */
  streakCount: number;
  /**
   * ISO date string of last activity (YYYY-MM-DD), or null
   * @nullable
   */
  lastActivityDate: string | null;
}

export interface StreakActivityRequest {
  clientId: string;
}

export interface StreakResult {
  /** New streak count after recording activity */
  streakCount: number;
  /** Whether the streak increased compared to the previous value */
  increased: boolean;
}

export interface GenerateIllustrationsRequest {
  /** Full story text */
  story: string;
  /** Story title */
  title: string;
  /** Child's name */
  childName: string;
  /** Story emoji */
  emoji: string;
}

export interface GenerateIllustrationsResponse {
  /** Base64-encoded PNG image data for each illustrated scene */
  images: string[];
}

export interface VoiceProfileResponse {
  /** ElevenLabs voice ID */
  voiceId: string;
  /** Display name for the cloned voice */
  voiceName: string;
}

export interface DeleteVoiceResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type GetStreakParams = {
  clientId: string;
};

export type GetSeriesParams = {
  childId: number;
};

export type GetVoiceParams = {
  clientId: string;
};

export type DeleteVoiceParams = {
  clientId: string;
};

export type GetMemoryParams = {
  childId: number;
  seriesId?: number;
};

export type DeleteMemoryParams = {
  childId: number;
  seriesId?: number;
};