# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 0.1.0
paths:
...
  /saved-stories:
    get:
      summary: Get all saved stories
      description: Returns all saved stories ordered by newest first
      responses:
        '200':
          description: List of saved stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedStory'
    post:
      summary: Save a story
      description: Saves a generated story to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveStoryRequest'
      responses:
        '201':
          description: Story saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedStory'
  /saved-stories/{id}:
    delete:
      summary: Delete a saved story
      description: Deletes a saved story by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
...
  /memory:
    get:
      summary: Get story memory for a child or series
      parameters:
        - name: childId
          in: query
          required: true
          schema:
            type: integer
        - name: seriesId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Story memory records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoryMemory'
    put:
      summary: Upsert story memory for a child or series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertMemoryRequest'
      responses:
        '200':
          description: Updated story memory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryMemory'
    delete:
      summary: Delete story memory for a child or series
      parameters:
        - name: childId
          in: query
          required: true
          schema:
            type: integer
        - name: seriesId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
components:
  schemas:
...
    SaveStoryRequest:
      type: object
      required: [childName, emoji, title, story, interests]
      properties:
        childName:
          type: string
        emoji:
          type: string
        title:
          type: string
        story:
          type: string
        storySummary:
          type: string
          description: Short 3-5 sentence summary for use in next episode
        interests:
          type: string
          description: Comma-separated interests
        childId:
          type: integer
          description: ID of the child profile to link this story to
        seriesId:
          type: integer
          description: ID of the series this story belongs to
        episodeNumber:
          type: integer
          description: Episode number within the series
    SavedStory:
      type: object
      required: [id, childName, emoji, title, story, interests, createdAt]
      properties:
        id:
          type: integer
        childName:
          type: string
        emoji:
          type: string
        title:
          type: string
        story:
          type: string
        storySummary:
          type: string
          nullable: true
          description: Short summary stored with the story
        interests:
          type: string
        createdAt:
          type: string
        childId:
          type: integer
          nullable: true
        seriesId:
          type: integer
          nullable: true
        episodeNumber:
          type: integer
          nullable: true
    MemoryCharacter:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    MemoryLocation:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    StoryMemory:
      type: object
      required: [id, userId, childId, mainCharacters, sideCharacters, locations, themes, tonePreferences, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        childId:
          type: integer
        seriesId:
          type: integer
          nullable: true
        mainCharacters:
          type: array
          items:
            $ref: '#/components/schemas/MemoryCharacter'
          description: Main recurring characters
        sideCharacters:
          type: array
          items:
            $ref: '#/components/schemas/MemoryCharacter'
          description: Supporting characters
        locations:
          type: array
          items:
            $ref: '#/components/schemas/MemoryLocation'
          description: Settings and locations
        themes:
          type: array
          items:
            type: string
          description: Recurring story themes
        tonePreferences:
          type: array
          items:
            type: string
          description: Observed tone preferences
        updatedAt:
          type: string
    UpsertMemoryRequest:
      type: object
      required: [childId]
      properties:
        childId:
          type: integer
        seriesId:
          type: integer
          nullable: true
        mainCharacters:
          type: array
          items:
            $ref: '#/components/schemas/MemoryCharacter'
        sideCharacters:
          type: array
          items:
            $ref: '#/components/schemas/MemoryCharacter'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/MemoryLocation'
        themes:
          type: array
          items:
            type: string
        tonePreferences:
          type: array
          items:
            type: string
...
    DeleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...