# lib/api-zod/src/generated/types/saveStoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface SaveStoryRequest {
  childName: string;
  emoji: string;
  title: string;
  story: string;
  /** Short 3-5 sentence summary for use in next episode */
  storySummary?: string;
  /** Comma-separated interests */
  interests: string;
  /** ID of the child profile to link this story to */
  childId?: number;
  /** ID of the series this story belongs to */
  seriesId?: number;
  /** Episode number within the series */
  episodeNumber?: number;
}