# lib/api-zod/src/generated/types/savedStory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface SavedStory {
  id: number;
  childName: string;
  emoji: string;
  title: string;
  story: string;
  /**
   * Short summary stored with the story
   * @nullable
   */
  storySummary?: string | null;
  interests: string;
  createdAt: string;
  /** @nullable */
  childId?: number | null;
  /** @nullable */
  seriesId?: number | null;
  /** @nullable */
  episodeNumber?: number | null;
}