# lib/api-zod/src/generated/types/storyMemory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryCharacter } from "./memoryCharacter";
import type { MemoryLocation } from "./memoryLocation";

export interface StoryMemory {
  id: number;
  userId: string;
  childId: number;
  /** @nullable */
  seriesId?: number | null;
  /** Main recurring characters */
  mainCharacters: MemoryCharacter[];
  /** Supporting characters */
  sideCharacters: MemoryCharacter[];
  /** Settings and locations */
  locations: MemoryLocation[];
  /** Recurring story themes */
  themes: string[];
  /** Observed tone preferences */
  tonePreferences: string[];
  updatedAt: string;
}