# lib/api-zod/src/generated/types/upsertMemoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryCharacter } from "./memoryCharacter";
import type { MemoryLocation } from "./memoryLocation";

export interface UpsertMemoryRequest {
  childId: number;
  /** @nullable */
  seriesId?: number | null;
  mainCharacters?: MemoryCharacter[];
  sideCharacters?: MemoryCharacter[];
  locations?: MemoryLocation[];
  themes?: string[];
  tonePreferences?: string[];
}